Break up no card display for clarity
